requirements.txt
torch>=0.4.1
torchvision>=0.2.1
tqdm>=4.23.4
visdom>=0.1.8
tensorflow-gpu
foolbox>=1.7.0